refactor: move setup internals behind core packages
