+++ docs/llms/coordination.md
@@ -169,7 +169,6 @@ Configure through provider setup plus `CoordinationOptions`.
 ### Dependencies
 
 - `Headless.Checks`
-- `FluentValidation`
 
 ### Side Effects
 
@@ -226,6 +225,7 @@ Set `HeartbeatInterval < SuspicionThreshold < DeadThreshold`; `DeadRetentionWind
 - `Headless.Core`
 - `Headless.Extensions`
 - `Headless.Hosting`
+- `FluentValidation`
 - `Microsoft.Extensions.Configuration.Abstractions`
 - `Microsoft.Extensions.Hosting.Abstractions`
 - `Microsoft.Extensions.Logging.Abstractions`
