+++ docs/llms/sql.md
@@ -1,6 +1,6 @@
 ---
 domain: SQL
-packages: Sql.Abstractions, Sql.PostgreSql, Sql.SqlServer, Sql.Sqlite
+packages: Sql.Abstractions, Sql.Core, Sql.PostgreSql, Sql.SqlServer, Sql.Sqlite
 ---
 
 # SQL
@@ -22,38 +22,45 @@ packages: Sql.Abstractions, Sql.PostgreSql, Sql.SqlServer, Sql.Sqlite
     - [Configuration](#configuration)
     - [Dependencies](#dependencies)
     - [Side Effects](#side-effects)
-- [Headless.Sql.PostgreSql](#headlesssqlpostgresql)
+- [Headless.Sql.Core](#headlesssqlcore)
     - [Problem Solved](#problem-solved-1)
     - [Key Features](#key-features-1)
-    - [Design Notes](#design-notes)
     - [Installation](#installation-1)
     - [Quick Start](#quick-start-1)
-    - [Configuration](#configuration-1)
     - [Dependencies](#dependencies-1)
     - [Side Effects](#side-effects-1)
-- [Headless.Sql.SqlServer](#headlesssqlsqlserver)
+- [Headless.Sql.PostgreSql](#headlesssqlpostgresql)
     - [Problem Solved](#problem-solved-2)
     - [Key Features](#key-features-2)
+    - [Design Notes](#design-notes)
     - [Installation](#installation-2)
     - [Quick Start](#quick-start-2)
-    - [Configuration](#configuration-2)
+    - [Configuration](#configuration-1)
     - [Dependencies](#dependencies-2)
     - [Side Effects](#side-effects-2)
-- [Headless.Sql.Sqlite](#headlesssqlsqlite)
+- [Headless.Sql.SqlServer](#headlesssqlsqlserver)
     - [Problem Solved](#problem-solved-3)
     - [Key Features](#key-features-3)
-    - [Design Notes](#design-notes-1)
     - [Installation](#installation-3)
     - [Quick Start](#quick-start-3)
-    - [Configuration](#configuration-3)
+    - [Configuration](#configuration-2)
     - [Dependencies](#dependencies-3)
     - [Side Effects](#side-effects-3)
+- [Headless.Sql.Sqlite](#headlesssqlsqlite)
+    - [Problem Solved](#problem-solved-4)
+    - [Key Features](#key-features-4)
+    - [Design Notes](#design-notes-1)
+    - [Installation](#installation-4)
+    - [Quick Start](#quick-start-4)
+    - [Configuration](#configuration-3)
+    - [Dependencies](#dependencies-4)
+    - [Side Effects](#side-effects-4)
 
 > Provider-agnostic SQL connection factory for raw SQL / Dapper scenarios with PostgreSQL, SQL Server, and SQLite backends.
 
 ## Quick Orientation
 
-Install `Headless.Sql.Abstractions` plus one provider package:
+Install `Headless.Sql.Abstractions` plus one provider package. Add `Headless.Sql.Core` when you need the default scoped `ISqlCurrentConnection` implementation:
 
 - `Headless.Sql.PostgreSql` — wraps Npgsql; returns `NpgsqlConnection`
 - `Headless.Sql.SqlServer` — wraps `Microsoft.Data.SqlClient`; returns `SqlConnection`
@@ -74,7 +81,7 @@ Inject `ISqlConnectionFactory` and call `CreateNewConnectionAsync()` to get an a
 - Do **not** construct connections directly (`new NpgsqlConnection(cs)` / `new SqlConnection(cs)`). Always go through the factory so the connection string is centralized and the factory can be swapped in tests.
 - Connections returned by `CreateNewConnectionAsync()` are **already open** — calling `OpenAsync()` on them again throws an `InvalidOperationException`.
 - Always dispose connections with `await using` — they are `IAsyncDisposable`. Holding an open connection unnecessarily may exhaust the connection pool.
-- `ISqlCurrentConnection` / `DefaultSqlCurrentConnection` provide an ambient, lazy-open connection for unit-of-work patterns. Inject as scoped, not singleton.
+- `ISqlCurrentConnection` defines an ambient, lazy-open connection for unit-of-work patterns. `Headless.Sql.Core` provides `DefaultSqlCurrentConnection`; register it as scoped, not singleton.
 - `IConnectionStringChecker` is for health checks and startup validation; register the provider implementation (e.g., `NpgsqlConnectionStringChecker`) and inject `IConnectionStringChecker`. Note: `SqliteConnectionStringChecker` always returns `DatabaseExists = true` when connected (SQLite creates the file on open).
 - For in-process integration tests, register `SqliteConnectionFactory` with `"Data Source=:memory:"` — it needs no external server.
 - There is no per-package `AddSql*()` extension. Manual `AddSingleton<ISqlConnectionFactory>(...)` registration is the only pattern.
@@ -106,7 +113,7 @@ public interface ISqlCurrentConnection : IAsyncDisposable
 }
 ```
 
-`DefaultSqlCurrentConnection` lazily opens one connection per instance, protects concurrent callers with an `AsyncLock`, and re-opens if the connection drops. Register it as **scoped** — one instance per request/scope — so it is disposed at the end of each unit of work.
+`Headless.Sql.Core` provides `DefaultSqlCurrentConnection`, which lazily opens one connection per instance, protects concurrent callers with an `AsyncLock`, and re-opens if the connection drops. Register it as **scoped** — one instance per request/scope — so it is disposed at the end of each unit of work.
 
 ### Connection string checker
 
@@ -137,7 +144,6 @@ Application code that works with raw SQL should not depend on a specific ADO.NET
 
 - `ISqlConnectionFactory` — create and manage database connections; `GetConnectionString()` retrieves the configured string; `CreateNewConnectionAsync()` returns an already-open `DbConnection`
 - `ISqlCurrentConnection` — ambient connection for unit-of-work scopes; lazy-opens on first call, re-opens on drop
-- `DefaultSqlCurrentConnection` — concrete thread-safe implementation of `ISqlCurrentConnection` backed by `AsyncLock`
 - `IConnectionStringChecker` — validate server reachability and database existence; returns `(bool Connected, bool DatabaseExists)`
 
 ### Installation
@@ -167,11 +173,7 @@ public sealed class OrderRepository(ISqlConnectionFactory connectionFactory)
 }
 ```
 
-Register `DefaultSqlCurrentConnection` as scoped when you need a shared ambient connection within a unit of work:
-
-```csharp
-builder.Services.AddScoped<ISqlCurrentConnection, DefaultSqlCurrentConnection>();
-```
+Add `Headless.Sql.Core` when you need the default scoped `ISqlCurrentConnection` implementation.
 
 ### Configuration
 
@@ -179,14 +181,51 @@ None. This is an abstractions-only package.
 
 ### Dependencies
 
-- `Headless.Hosting`
-- `Nito.AsyncEx` (transitively, via `DefaultSqlCurrentConnection`)
+None.
 
 ### Side Effects
 
 None. This is an abstractions package — it registers no services.
 
 ---
+
+## Headless.Sql.Core
+
+Default implementation package for provider-agnostic SQL helpers.
+
+### Problem Solved
+
+Keeps `Headless.Sql.Abstractions` limited to interfaces while providing a reusable scoped ambient connection implementation for unit-of-work patterns.
+
+### Key Features
+
+- `DefaultSqlCurrentConnection` — concrete thread-safe implementation of `ISqlCurrentConnection` backed by `AsyncLock`.
+- Lazily opens one connection per scope and reuses it until disposal.
+- Reopens the underlying connection if it is observed closed.
+
+### Installation
+
+```bash
+dotnet add package Headless.Sql.Core
+```
+
+### Quick Start
+
+```csharp
+builder.Services.AddScoped<ISqlCurrentConnection, DefaultSqlCurrentConnection>();
+```
+
+### Dependencies
+
+- `Headless.Sql.Abstractions`
+- `Nito.AsyncEx`
+
+### Side Effects
+
+None. Register services explicitly.
+
+---
+
 ## Headless.Sql.PostgreSql
 
 PostgreSQL connection factory backed by Npgsql.
@@ -255,7 +294,9 @@ services.AddSingleton<ISqlConnectionFactory>(sp =>
 
 ### Dependencies
 
+- `Headless.Checks`
 - `Headless.Sql.Abstractions`
+- `Microsoft.Extensions.Logging.Abstractions`
 - `Npgsql`
 
 ### Side Effects
@@ -327,6 +368,7 @@ services.AddSingleton<ISqlConnectionFactory>(sp =>
 
 - `Headless.Sql.Abstractions`
 - `Microsoft.Data.SqlClient`
+- `Microsoft.Extensions.Logging.Abstractions`
 
 ### Side Effects
 
@@ -396,6 +438,7 @@ Pass the connection string directly to the constructor. SQLite connection string
 
 - `Headless.Sql.Abstractions`
 - `Microsoft.Data.Sqlite`
+- `Microsoft.Extensions.Logging.Abstractions`
 
 ### Side Effects
