+++ src/Headless.Api.Mvc/Headless.Api.Mvc.csproj
@@ -7,7 +7,6 @@
     <PackageReference Include="Asp.Versioning.Mvc" />
     <PackageReference Include="Asp.Versioning.Mvc.ApiExplorer" />
     <PackageReference Include="Mediator.Abstractions" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Headless.Api.Core\Headless.Api.Core.csproj" />
