+++ src/Headless.AuditLog.Abstractions/Headless.AuditLog.Abstractions.csproj
@@ -4,6 +4,7 @@
     <RootNamespace>Headless.AuditLog</RootNamespace>
   </PropertyGroup>
   <ItemGroup>
+    <InternalsVisibleTo Include="Headless.AuditLog.Core" />
     <InternalsVisibleTo Include="Headless.AuditLog.Storage.EntityFramework" />
     <InternalsVisibleTo Include="Headless.AuditLog.Storage.PostgreSql" />
     <InternalsVisibleTo Include="Headless.AuditLog.Storage.SqlServer" />
@@ -11,10 +12,5 @@
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Headless.Extensions\Headless.Extensions.csproj" />
-    <ProjectReference Include="..\Headless.Hosting\Headless.Hosting.csproj" />
-  </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" />
-    <PackageReference Include="Microsoft.Extensions.Options" />
   </ItemGroup>
 </Project>
