+++ src/Headless.AuditLog.Core/HeadlessAuditLogBuilder.cs
