+++ src/Headless.AuditLog.Core/HeadlessAuditLogSetupBuilder.cs
