+++ src/Headless.AuditLog.Core/IAuditLogStorageOptionsExtension.cs
