+++ src/Headless.AuditLog.Core/Setup.cs
