+++ src/Headless.AuditLog.Storage.EntityFramework/Headless.AuditLog.Storage.EntityFramework.csproj
@@ -13,6 +13,7 @@
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Headless.AuditLog.Abstractions\Headless.AuditLog.Abstractions.csproj" />
+    <ProjectReference Include="..\Headless.AuditLog.Core\Headless.AuditLog.Core.csproj" />
     <ProjectReference Include="..\Headless.Orm.EntityFramework\Headless.Orm.EntityFramework.csproj" />
   </ItemGroup>
 </Project>
