+++ src/Headless.Coordination.Core/Headless.Coordination.Core.csproj
@@ -13,6 +13,7 @@
     <InternalsVisibleTo Include="Headless.Coordination.Core.Tests.Unit" />
   </ItemGroup>
   <ItemGroup>
+    <PackageReference Include="FluentValidation" />
     <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" />
     <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" />
     <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" />
