+++ src/Headless.Sql.Abstractions/Headless.Sql.Abstractions.csproj
@@ -3,7 +3,4 @@
     <TargetFramework>net10.0</TargetFramework>
     <RootNamespace>Headless.Sql</RootNamespace>
   </PropertyGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\Headless.Hosting\Headless.Hosting.csproj" />
-  </ItemGroup>
 </Project>
