+++ src/Headless.Sql.Core/Headless.Sql.Core.csproj
@@ -1,10 +1,10 @@
-﻿<Project Sdk="Headless.NET.Sdk">
+<Project Sdk="Headless.NET.Sdk">
   <PropertyGroup>
     <TargetFramework>net10.0</TargetFramework>
+    <RootNamespace>Headless.Sql</RootNamespace>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Microsoft.Data.SqlClient" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" />
+    <PackageReference Include="Nito.AsyncEx" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Headless.Sql.Abstractions\Headless.Sql.Abstractions.csproj" />
