+++ tests/Headless.AuditLog.Core.Tests.Unit/Headless.AuditLog.Core.Tests.Unit.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Headless.NET.Sdk.Test">
+  <PropertyGroup>
+    <TargetFramework>net10.0</TargetFramework>
+    <RootNamespace>Tests</RootNamespace>
+    <OutputType>Exe</OutputType>
+  </PropertyGroup>
+  <ItemGroup>
+    <PackageReference Include="xunit.v3.mtp-v2" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\Headless.AuditLog.Core\Headless.AuditLog.Core.csproj" />
+  </ItemGroup>
+</Project>
