+++ tests/Headless.Coordination.Core.Tests.Unit/CoordinationOptionsValidatorTests.cs
