+++ tests/Headless.Sql.Core.Tests.Unit/DefaultSqlCurrentConnectionTests.cs
