+++ tests/Headless.Sql.Tests.Harness/Headless.Sql.Tests.Harness.csproj
@@ -7,6 +7,7 @@
   </PropertyGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\src\Headless.Sql.Abstractions\Headless.Sql.Abstractions.csproj" />
+    <ProjectReference Include="..\..\src\Headless.Sql.Core\Headless.Sql.Core.csproj" />
     <ProjectReference Include="..\..\src\Headless.Testing\Headless.Testing.csproj" />
   </ItemGroup>
 </Project>
